backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RedactIQ Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.0.post1",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.2.2",
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "presidio-analyzer>=2.2.351",
    "spacy>=3.7.2",
    "pytesseract>=0.3.10",
    "Pillow>=10.2.0",
]

[tool.uv]
managed = true
package = false
